backend/utils/ocr_functions.py
import cv2
import numpy as np
import os
import logging

logger = logging.getLogger(__name__)

def detect_code_box(image_path):
    """
    Detect exam code from code box image
    TODO: Implement actual OCR logic
    """
    try:
        # Mock implementation - replace with actual OCR
        return "101"
    except Exception as e:
        logger.error(f"Error detecting code box: {e}")
        return None

def detect_name_student(image_path, student_names):
    """
    Detect student name from name box image
    TODO: Implement actual OCR logic
    """
    try:
        # Mock implementation - replace with actual OCR
        return "Nguyễn Văn A"
    except Exception as e:
        logger.error(f"Error detecting student name: {e}")
        return None

def detect_id_student(image_path, student_ids):
    """
    Detect student ID from ID box image
    TODO: Implement actual OCR logic
    """
    try:
        # Mock implementation - replace with actual OCR
        return "SV001"
    except Exception as e:
        logger.error(f"Error detecting student ID: {e}")
        return None

def detect_index_student(image_path):
    """
    Detect student index from index box image
    TODO: Implement actual OCR logic
    """
    try:
        # Mock implementation - replace with actual OCR
        return "1"
    except Exception as e:
        logger.error(f"Error detecting student index: {e}")
        return None

def predict_grade(grading_path):
    """
    Predict grade from grading table image
    TODO: Implement actual grading logic
    """
    try:
        # Mock implementation - replace with actual grading
        answers = {1: 'A', 2: 'B', 3: 'C', 4: 'D', 5: 'A', 6: 'B', 7: 'C', 8: 'D'}
        return grading_path, answers
    except Exception as e:
        logger.error(f"Error predicting grade: {e}")
        return grading_path, None

def calculate_score(answers, df_key, exam_code):
    """
    Calculate score based on answers and answer key
    TODO: Implement actual scoring logic
    """
    try:
        # Mock implementation - replace with actual scoring
        return 8.0
    except Exception as e:
        logger.error(f"Error calculating score: {e}")
        return 0.0

def overlay_image(original_path, overlay_path, grading_box, output_path):
    """
    Overlay grading result on original image
    TODO: Implement actual overlay logic
    """
    try:
        # Mock implementation - replace with actual overlay
        original_img = cv2.imread(original_path)
        if original_img is None:
            return False
        
        cv2.imwrite(output_path, original_img)
        return True
    except Exception as e:
        logger.error(f"Error overlaying image: {e}")
        return False

def copy_to_static(file_path, temp_file_name):
    """
    Copy file to static directory
    TODO: Implement actual file copying
    """
    try:
        # Mock implementation - replace with actual file copying
        return file_path
    except Exception as e:
        logger.error(f"Error copying to static: {e}")
        return file_path 